feat(useObservable): now returns destructured values, including an `onError` hook
Per a [suggestion here](https://github.com/vueuse/vueuse/issues/566#issuecomment-857228255) this PR attempts to use an `onError` "hook" pattern. However it should be noted that this will be a breaking change.

Resolves #566

BREAKING CHANGE: The return value of `useObservable` now returns a destructured object containing the ref as `out`, and an `onError` hook registration point.