added useFocusInputElement
added a function to focus the input element found within the specified ref.
This is needed when using UI libraries that wrap the input elements in
layers of divs, making it difficult to get a ref to the actual input element
in the client code. This function searchs the specified ref for an input
element and then uses useFocus with that found input element.

I also fixed up the line endings of useFocus and the demo for useFocus
and a compile error in useFocus.