Added the notion of isSupported ('Notification' in window).
Added Added the notion of isSupported ('Notification' in window). Use close() to remove a notification that is no longer relevant to to the user (e.g.the user already read the notification on the webpage). Added a notificationClose() method for manual triggering of closing the notification if needed. Updated index.md usage example to reference the new isSupported boolean return. Includes fixing minor typos in TSDoc.