fix(unrefElement): don't return the Vue instance when `$el` is `null`/`undefined`
When using unrefElement on a template ref of a Vue component which is not available (`v-if="false"`) it returns the Vue instance object with `{ $el: null }`. It should return `null` instead of the Vue instance object.

The cause of this bug is how `plain` is checked to be a VueInstance. Rather than using the value of `$el`, it should check if it has a property named `$el`.