fix(useScroll): account for rounding errors when calculating arrivedState
Sometimes when zooming the page to sizes other than 100% and scrolling a
container all the way to the bottom right, the calculation for
`arrivedState.right` and `arrivedState.bottom` would incorrectly say
`false`. This is due to a [rounding error][] because `scrollTop` is a
non-rounded number, while `scrollHeight` and `clientHeight` are rounded.

Using the MDN documentation on `Element.scrollHeight` as guidance,
this commit factors in a 1px margin of error to more accurately
calculate `arrivedState.right` and `arrivedState.bottom`.

[rounding error]: https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollHeight#determine_if_an_element_has_been_totally_scrolled

fix #2047