feat (core): support shadow roots in useActiveElement
Assuming you have a DOM structure like so:

```
body
  my-element
    #shadow-root
      input
```

When the `input` becomes active, the outer document's `activeElement`
will actually be `my-element`, _not_ the input.

Each shadow root has its own `activeElement` in case you want to get
hold of the inner element in these cases.

This adds support to `useActiveElement` such that you can pass a shadow
root as the document to observe the active element of.