fix(useStorage): ensure setting value to null syncs to other instances
Previously, when using `useStorage` in multiple components, setting one
of their values to `null` would cause the storage item to be deleted and
the returned ref to be correct in that component, but in other
components the item would still be set to its old value.

I've had some trouble with the tests - one of the other tests has
started failing when all the tests are ran together, but passes when ran
in isolation with `.only`. Not sure why - could use some advise here!