chore: upgrade playwright
CI is failing at the minute because `nlx playwright install` will
install the _latest_ playwright browsers. These then mismatch the
version of playwright we have installed (which tries to resolve older
browser versions).

For now, this upgrades playwright to use the same version.

Later, we should investigate how to constrain `nlx playwright install`
to the same version so it has to be manually updated in future.