fix: always mount listeners in useStorage
This is _yet another_ possible fix for the various `useStorage` scope
issues.

Basically, `onMounted` will _always_ run in the scope of the
_component_. This means the event listener will not be torn down unless
the element itself is. When the various effect scopes we use are torn
down, the listener will remain and cause various memory leaks.

We can solve this by moving the event listeners _outside_ the
`onMounted` callback, as they will then teardown when the scope they're
in disposes.

However, this then means the event listener could run before the
component has mounted, when `initOnMount` is true. To solve this, we
check in the event handler and do nothing if it isn't mounted yet.

We only care about the _first_ mount, since all others will have already
gone through at least one `update` call.