feat(useAsyncState): allow initial value to be a ref
Outside of typescript, we can already do this but we cause a maximum
stack call exception since we're setting `ref.value = ref` (infinitely).

This updates the source so you can pass a ref in and we turn it into a
value before using it as such.

This means you can now use your own ref:

```ts
const myRef = shallowRef(200);
useAsyncState(asyncPromise, myRef);
// some time later, `myRef.value` is whatever `asyncPromise` resolved to
```