feat(useDropZone): add checkValidity function
This change adds a new `checkValidity` option for custom validation of dropped items to give users more control over the items allowed to drop than only `dataTypes`.

For me, this fixes this very concrete problem:

In my app, there is an editor, where users can drag text around. Having a global dropzone to upload files will also intercept this text dragging and dropping. For `dataTypes`, text that's dragged around from an editor also has a type of `text/plain`, the same as a `.txt` file. This makes it impossible to distinguish between a text file (which is a legitimate upload file) and text the user is dragging around in the editor.
Exposing the `DataTransferItemList` for filtering allows me to use the `kind` property of one of these items to filter the dropzone for files only.